pass argv to init function for main set macro
aesop_main_set_with_init() allows the caller to specify a function to
run before calling triton_init() (for example, to set zeroconf
parameters before starting triton).  This is a small tweak to pass argc
and argv to that function in case there are command line arguments to
handle before calling triton_init().

ref #81