Fix ae_cancel_ctl issue
pbranch calling blocking function does not indicate the blocking
function as a child but instead it uses a 0 for resourceid and the
child ctl ptr is stored in the data field of the opid.

Fixed ae_cancel_ctl to properly recurse.

This fixes an issue where cancellation would suddenly stop working if a
blocking function called the resource function instead of calling the
resource function directly.